+++ docs/release_notes.rst
@@ -2,6 +2,11 @@
 
 Release Notes
 =============
+3.6.6
+------
+* Bugfix: Label confusion with curve maths for curves greater than 26
+* Refactoring `gety`, `getymax`, `getymin`, `max_curve`, `min_curve`, and `average` to use `numpy.interp()`
+
 3.6.5
 ------
 * Bugfix: Launcher update for HPCs read files from command line